Migrate away from floats because they're bad

+++ Game.h
@@ -28,7 +28,7 @@ class Game {
         void drawEntity(Entity * entity);
     
         void renderPlayer(Player * player);
-        void checkCollisions(std::pair<float,float> coords);
+        void checkCollisions(std::pair<int,int> coords);
         void drawCollisionbox(int boxX, int boxY);
         void drawTileBox(int tileIndex);
         void clean();
